requirements/prod.txt
setuptools
mutagen
simplejson
# Modern Google Cloud Datastore client library
google-cloud-datastore
# Required for service account impersonation
google-auth
# Legacy: used by App Engine SDK (can be removed after migration complete)
pycryptodome
lxml
